resume-parser-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
PyMuPDF==1.23.8
python-docx==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
bcrypt==4.1.2
